requirements.txt
accelerate==1.9.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
backoff==2.2.1
beautifulsoup4==4.13.4
bitsandbytes==0.46.1
bs4==0.0.2
build==1.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.3
ddgs==9.5.1
debugpy==1.8.15
decorator==5.2.1
duckduckgo_search==8.1.1
executing==2.2.0
filelock==3.18.0
fsspec==2025.5.1
googleapis-common-protos==1.70.0
googlesearch-python==1.3.0
grpcio==1.74.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
installer==0.7.0
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
langfuse==3.2.2
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openinference-instrumentation==0.1.36
openinference-instrumentation-smolagents==0.1.14
openinference-semantic-conventions==0.1.21
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
portalocker==3.2.0
primp==0.15.0
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
pyzmq==27.0.0
qdrant-client==1.15.1
regex==2025.7.34
requests==2.32.4
rich==14.1.0
safetensors==0.5.3
sentencepiece==0.2.0
setuptools==80.9.0
six==1.17.0
smolagents @ git+https://github.com/huggingface/smolagents.git@f7b4a33862cf5459b060f3865cd267458bfc3188
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
sympy==1.14.0
tokenizers==0.21.4
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.1
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
wheel==0.45.1
wrapt==1.17.2
zipp==3.23.0
